Allow ValidationErrors type to be used externally
The `ValidationErrors` type is public, but inside its 2 functions it casts to the private `fieldError`. Since there is no way to access the `fieldError` outside this package it meant creating and using `ValidationErrors` outside this package was impossible.

My use case is for testing, I'm trying to test how my code behaves when various `ValidationErrors` are returned without needing to actually setup data to fail in the various ways.