Expose the struct the field is contained within
This allows users of `FieldError` to be able to query information about
the struct the field is contained within. This information can then be
used to work out what to display when the param field points to another
field, such as with `gtfield`, and a tag name function has been
registered.