Omitzero does not work with slice and map bug (#1436)
## Fixes 

Fixed a bug where using omitzero still validated empty structs and maps.
This PR ensures that if omitzero is used as a tag then empty slices and
maps are not validated.

* Modified the `baked_in.go` file to ensure that slices and maps are
considered non zero only when they are not nil and they are not empty.
* Added tests to ensure empty slices and maps are not validated while
using omitzero tag.

Check #1424 for more details.

**Make sure that you've checked the boxes below before you submit PR:**
- [x] Tests exist or have been written that cover this particular
change.

@go-playground/validator-maintainers