Fix: Validator panics when 'nil' is used along with required if for slices and maps (#1442)
## Fixes 
Added support to check for 'nil' values in case of maps and slices when
used with required if. (Previously only lengths were being checked as a
result using nil with required if would lead to a panic since the
program tried to convert 'nil' to an integer)
Created a separate case for arrays to only check lengths since arrays
cannot be nil in go.

Check #1437 for more details.

**Make sure that you've checked the boxes below before you submit PR:**
- [x] Tests exist or have been written that cover this particular
change.

@go-playground/validator-maintainers